Ensure validation passes
