+++ bish/command/planner.go
@@ -19,6 +19,7 @@ Grammar:
 			 |- < String >
 */
 
+// The ExecutionPlanner parses a tokenized sequence and plans command execution as an executable tree.
 type ExecutionPlanner struct {
 	Args  []string
 	ctx   *state.State
@@ -30,6 +31,7 @@ type ExecutionPlanner struct {
 	done       bool
 }
 
+// NewExecutionPlanner returns a planner from the shell state and a sequence of arguments.
 func NewExecutionPlanner(ctx *state.State, args []string) *ExecutionPlanner {
 	e := &ExecutionPlanner{Args: args, ctx: ctx}
 
@@ -56,6 +58,7 @@ func (p *ExecutionPlanner) accept(char string) bool {
 	return false
 }
 
+// Argument parses an argument from the current state.
 func (p *ExecutionPlanner) Argument() (Argument, error) {
 	if p.accept("(") {
 		subcmd, err := p.Command(false)
@@ -94,7 +97,7 @@ func (p *ExecutionPlanner) Command(topLevel bool) (Command, error) {
 		argumentBuffer = append(argumentBuffer, arg)
 
 		if p.done || *p.nextTok == ")" || *p.nextTok == ">" {
-			baseCmd := &CommandTree{
+			baseCmd := &Executable{
 				Args: argumentBuffer,
 				Ctx:  p.ctx,
 			}
@@ -103,7 +106,7 @@ func (p *ExecutionPlanner) Command(topLevel bool) (Command, error) {
 			}
 			return baseCmd, nil
 		} else if p.accept("|") {
-			aCmd := &CommandTree{
+			aCmd := &Executable{
 				Args:  argumentBuffer,
 				Shell: true,
 				Ctx:   p.ctx,
