+++ bish/hooks/hook.go
@@ -1,5 +1,6 @@
 package hooks
 
+// A Hook pre-processes command arguments.
 type Hook interface {
 	Apply(args []string) []string
 }
