+++ bish/state/state.go
@@ -13,17 +13,19 @@ type State struct {
 	mtx   sync.RWMutex
 }
 
-// NewContext returns a new shell context store.
+// New returns a new shell context store.
 func New() *State {
 	return &State{data: make(map[string]interface{}), hooks: make(map[string]hooks.Hook)}
 }
 
+// AddHook registers a new hook in shell state.
 func (c *State) AddHook(name string, hk hooks.Hook) {
 	c.mtx.Lock()
 	defer c.mtx.Unlock()
 	c.hooks[name] = hk
 }
 
+// Hooks returns all currently-defined hooks.
 func (c *State) Hooks() []hooks.Hook {
 	var hks []hooks.Hook
 	for _, v := range c.hooks {
@@ -42,6 +44,7 @@ func (c *State) GetKey(domain string, key string) (interface{}, bool) {
 	return v, ok
 }
 
+// SetKey sets a key in the context store.
 func (c *State) SetKey(domain string, key string, value interface{}) {
 	c.mtx.Lock()
 	defer c.mtx.Unlock()
