+++ cmd/bish/main.go
@@ -1,66 +1,19 @@
 package main
 
-import (
-	"io"
-	"os"
+import "github.com/dalloriam/bish/bish"
 
-	"github.com/chzyer/readline"
-	"github.com/dalloriam/bish/bish"
-)
-
-type TerminalBackend struct {
-	rl *readline.Instance
-}
-
-func newBackend() (*TerminalBackend, error) {
-	rl, err := readline.New("➤ ")
-	if err != nil {
-		return nil, err
-	}
-
-	return &TerminalBackend{rl: rl}, nil
-}
-
-func (t *TerminalBackend) UpdatePrompt(prompt string) {
-	t.rl.SetPrompt(prompt)
-}
-
-func (t *TerminalBackend) SetConfig(c readline.AutoCompleter) {
-	t.rl.Config.AutoComplete = c
-}
-
-func (t *TerminalBackend) Stderr() io.Writer {
-	return os.Stderr
-}
-
-func (t *TerminalBackend) Stdin() io.Reader {
-	return os.Stdin
-}
-
-func (t *TerminalBackend) Stdout() io.Writer {
-	return os.Stdout
-}
-
-func (t *TerminalBackend) ReadLine() (string, error) {
-	return t.rl.Readline()
-}
-
-func (t *TerminalBackend) Close() error {
-	return t.rl.Close()
-}
-
-func shellStart() {
+func main() {
 	backend, err := newBackend()
 	if err != nil {
 		panic(err)
 	}
-	defer backend.Close()
+	defer func() {
+		if err := backend.Close(); err != nil {
+			panic(err)
+		}
+	}()
 
 	shell := bish.New(backend)
 	backend.SetConfig(shell.CompletionProvider)
 	shell.Start()
-}
-
-func main() {
-	shellStart()
 }
